+++ solution/day04/src/test/article.test.ts
@@ -46,6 +46,25 @@ describe('Article', () => {
     );
   });
 
+  test('it should add a comment with the provided date', () => {
+    const article = new Article(
+      'Lorem Ipsum',
+      'consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore'
+    );
+    const now = new Date(2020, 0, 12);
+    jest.setSystemTime(now);
+    const creationDate = new Date(2022, 5, 20);
+
+    article.addComment('Amazing article !!!', 'Pablo Escobar', creationDate);
+
+    expect(article.getComments()).toHaveLength(1);
+    expect(article.getComments()).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({ creationDate })
+      ])
+    );
+  });
+
   test('it should throw an exception when adding existing comment', () => {
     const article = new Article(
       'Lorem Ipsum',