+++ solution/day04/src/CHANGELOG.md
@@ -1,3 +1,4 @@
 Improvements to the tests:
 1. "it should add valid comment" - no assertion; added the missing assertion by combining this and the following two tests: "it should add a comment with the given text", "it should add a comment with the given author"
 2. "it should add a comment with the date of the day" - no assertion, added the missing one.
+3. Added a test that checks if the provided comment date is saved.